Changed version

+++ mod/assign/feedback/editpdf/classes/document_services.php
@@ -166,7 +166,7 @@ EOD;
 
         // Capability checks.
         if (!$assignment->can_view_submission($userid)) {
-            throw new \moodle_exception('nopermission');
+            print_error('nopermission');
         }
 
         $files = array();
@@ -278,7 +278,7 @@ EOD;
 
         // Capability checks.
         if (!$assignment->can_view_submission($userid)) {
-            throw new \moodle_exception('nopermissiontoaccesspage', 'error');
+            print_error('nopermission');
         }
 
         $grade = $assignment->get_user_grade($userid, true, $attemptnumber);
@@ -363,7 +363,7 @@ EOD;
         $assignment = self::get_assignment_from_param($assignment);
 
         if (!$assignment->can_view_submission($userid)) {
-            throw new \moodle_exception('nopermission');
+            print_error('nopermission');
         }
 
         // When in readonly we can return the number of images in the DB because they should already exist,
@@ -400,7 +400,7 @@ EOD;
         $assignment = self::get_assignment_from_param($assignment);
 
         if (!$assignment->can_view_submission($userid)) {
-            throw new \moodle_exception('nopermission');
+            print_error('nopermission');
         }
 
         // Need to generate the page images - first get a combined pdf.
@@ -408,7 +408,7 @@ EOD;
 
         $status = $document->get_status();
         if ($status === combined_document::STATUS_FAILED) {
-            throw new \moodle_exception('Could not generate combined pdf.');
+            print_error('Could not generate combined pdf.');
         } else if ($status === combined_document::STATUS_PENDING_INPUT) {
             // The conversion is still in progress.
             return [];
@@ -438,13 +438,9 @@ EOD;
         $fs->delete_area_files($record->contextid, $record->component, $record->filearea, $record->itemid);
 
         $files = array();
-        $images = $pdf->get_images();
         for ($i = 0; $i < $pagecount; $i++) {
             try {
-                if (empty($images[$i])) {
-                    throw new \moodle_exception('error image');
-                }
-                $image = $images[$i];
+                $image = $pdf->get_image($i);
                 if (!$resetrotation) {
                     $pagerotation = page_editor::get_page_rotation($grade->id, $i);
                     $degree = !empty($pagerotation) ? $pagerotation->degree : 0;
@@ -503,7 +499,7 @@ EOD;
         $assignment = self::get_assignment_from_param($assignment);
 
         if (!$assignment->can_view_submission($userid)) {
-            throw new \moodle_exception('nopermission');
+            print_error('nopermission');
         }
 
         if ($assignment->get_instance()->teamsubmission) {
@@ -580,9 +576,7 @@ EOD;
             }
         }
 
-        $totalpagesforattempt = self::page_number_for_attempt($assignment, $userid, $attemptnumber, false);
-        // Here we are comparing the total number of images against the total number of pages from the combined PDF.
-        if (empty($pages) || count($pages) != $totalpagesforattempt) {
+        if (empty($pages)) {
             if ($readonly) {
                 // This should never happen, there should be a version of the pages available
                 // whenever we are requesting the readonly version.
@@ -649,10 +643,10 @@ EOD;
         $assignment = self::get_assignment_from_param($assignment);
 
         if (!$assignment->can_view_submission($userid)) {
-            throw new \moodle_exception('nopermission');
+            print_error('nopermission');
         }
         if (!$assignment->can_grade()) {
-            throw new \moodle_exception('nopermission');
+            print_error('nopermission');
         }
 
         // Need to generate the page images - first get a combined pdf.
@@ -660,7 +654,7 @@ EOD;
 
         $status = $document->get_status();
         if ($status === combined_document::STATUS_FAILED) {
-            throw new \moodle_exception('Could not generate combined pdf.');
+            print_error('Could not generate combined pdf.');
         } else if ($status === combined_document::STATUS_PENDING_INPUT) {
             // The conversion is still in progress.
             return false;
@@ -818,7 +812,7 @@ EOD;
         $assignment = self::get_assignment_from_param($assignment);
 
         if (!$assignment->can_view_submission($userid)) {
-            throw new \moodle_exception('nopermission');
+            print_error('nopermission');
         }
 
         $grade = $assignment->get_user_grade($userid, true, $attemptnumber);
@@ -854,10 +848,10 @@ EOD;
         $assignment = self::get_assignment_from_param($assignment);
 
         if (!$assignment->can_view_submission($userid)) {
-            throw new \moodle_exception('nopermission');
+            print_error('nopermission');
         }
         if (!$assignment->can_grade()) {
-            throw new \moodle_exception('nopermission');
+            print_error('nopermission');
         }
 
         $grade = $assignment->get_user_grade($userid, true, $attemptnumber);
@@ -920,13 +914,10 @@ EOD;
         $oldfile = $fs->get_file($record->contextid, $record->component, $record->filearea,
             $record->itemid, $record->filepath, $record->filename);
 
-        if ($oldfile) {
-            $newhash = \file_storage::hash_from_path($newfilepath);
-            if ($newhash === $oldfile->get_contenthash()) {
-                // Use existing file if contenthash match.
-                return $oldfile;
-            }
-            // Delete existing file.
+        $newhash = sha1($newfilepath);
+
+        // Delete old file if exists.
+        if ($oldfile && $newhash !== $oldfile->get_contenthash()) {
             $oldfile->delete();
         }
 
@@ -951,7 +942,7 @@ EOD;
         $grade = $assignment->get_user_grade($userid, true, $attemptnumber);
         // Check permission.
         if (!$assignment->can_view_submission($userid)) {
-            throw new \moodle_exception('nopermission');
+            print_error('nopermission');
         }
 
         $filearea = self::PAGE_IMAGE_FILEAREA;
